So about that 0.1 release...
This makes directory monitoring stable. The logging is a little bit too verbose for the info level; that will get fixed soon. Unfortunately, I can't do anything about the overflows that cause file change notifications to be lost.

The idea I had with generating hash codes to prevent duplicate insertions was okay, but I think it needs some more optimization so that a boatload of find operations aren't done.