It "works"! I think.
This commit adds a quick and dirty actor to help deal with
file modifications as another program is writing out the
compressed input for this program. There's a mechanism to
prevent multiple actors from working on the same file, and
a five-second retry if there's an I/O error when reading
the compressed input.